requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Login==0.6.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
packaging==24.2
python-dotenv==1.0.1
setuptools==75.8.0
Werkzeug==3.1.3
